[fix] 장바구니 수정

+++ src/pages/user/Cart/CartItem.js
@@ -6,6 +6,8 @@ import { useState,useEffect } from 'react';
 import '../../../styles/fonts.css';
 import styled from 'styled-components';
 import Checkbox from '@mui/material/Checkbox';
+import { DELETE_CartItem } from '@api/cart';
+
 
 const CartItem = ({ item, updateQuantity, updateChecked, deleteItem  }) => {
   const [quantity, setQuantity] = useState(item.quantity);
@@ -41,7 +43,16 @@ const CartItem = ({ item, updateQuantity, updateChecked, deleteItem  }) => {
   };
 
   const handleDelete = () => {
-    deleteItem(item.id);
+    DELETE_CartItem(item.id) // API 요청을 통해 상품 삭제
+      .then((response) => {
+        // 삭제 성공 시 deleteItem 함수 호출하여 UI 업데이트
+        if (response.success) {
+          deleteItem(item.id);
+        }
+      })
+      .catch((error) => {
+        console.log(error);
+      });
   };
